Move SharedObserver into observe crate
This change moves the SharedObserver into the observe crate, where it
fits in more naturally now. It also restructures the observe crate a
little, such that it does not become a one-file-dumping-ground.
