+++ rust/template/observe/subscription.rs
@@ -1,0 +1,6 @@
+/// A trait representing an object subscribed to an `Observable`.
+pub trait Subscription {
+    /// Cancel the subscription so the corresponding Observer
+    /// stops listening to the corresponding Observable.
+    fn unsubscribe(self: Box<Self>);
+}
