+++ src/Language/DifferentialDatalog/Compile.hs
@@ -160,6 +160,9 @@ rustLibFiles specname =
         , (dir </> "cmd_parser/parse.rs"                , $(embedFile "rust/template/cmd_parser/parse.rs"))
         , (dir </> "observe/Cargo.toml"                 , $(embedFile "rust/template/observe/Cargo.toml"))
         , (dir </> "observe/lib.rs"                     , $(embedFile "rust/template/observe/lib.rs"))
+        , (dir </> "observe/observable.rs"              , $(embedFile "rust/template/observe/observable.rs"))
+        , (dir </> "observe/observer.rs"                , $(embedFile "rust/template/observe/observer.rs"))
+        , (dir </> "observe/subscription.rs"            , $(embedFile "rust/template/observe/subscription.rs"))
         , (dir </> "ovsdb/Cargo.toml"                   , $(embedFile "rust/template/ovsdb/Cargo.toml"))
         , (dir </> "ovsdb/lib.rs"                       , $(embedFile "rust/template/ovsdb/lib.rs"))
         , (dir </> "ovsdb/test.rs"                      , $(embedFile "rust/template/ovsdb/test.rs"))
